+++ internal/dto/services/get.go
@@ -1,10 +1,10 @@
 package dto
 
 type GetServiceReq struct {
-	ServiceID int64 `validate:"required"`
+	ServiceID int32 `validate:"required"`
 }
 
 type GetServiceRes struct {
-	ServiceID   int64
+	ServiceID   int32
 	ServiceName string
 }
